[X] 파일 추가, 정리
