+++ PhotoSurfer-iOS/PhotoSurfer-iOS/Source/Presentation/HomeSearch/Views/HomeSearchHeaderCollectionReusableView.swift
@@ -1,0 +1,17 @@
+//
+//  HomeSearchHeaderCollectionReusableView.swift
+//  PhotoSurfer-iOS
+//
+//  Created by 김혜수 on 2022/07/15.
+//
+
+import UIKit
+
+class HomeSearchHeaderCollectionReusableView: UICollectionReusableView {
+
+    override func awakeFromNib() {
+        super.awakeFromNib()
+        // Initialization code
+    }
+    
+}
